requirements.txt
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
matplotlib==3.2.2
numpy==1.19.0
opencv-python==4.2.0.34
pyprog==1.1.0.post0
scikit-image==0.16.2
scikit-learn==0.23.1
scipy==1.4.1
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.0
tensorflow-probability==0.9.0
